Moved edit storyline to UI/CRUD. Also added storylines folder and moved createStoryline there. Also added a clear in add storyline when choosing exit.
